Notebooks and final mods
